Fix reversed byte order in custom file protocol
Before, the packed integer was being turned into a reversed byte array. This isn't really an issue as long as it's converted back into a packed integer using the same logic.

`settings.web-directory.read-only` will need to be set to `false` in the config when updating Pl3xMap to the version with this fix, since it includes changes to the webmap files.